moved analyticsInterceptingImage class to utils - added cypress/fixtures folder to .gitignore - removed added wait time for page load
