+++ cypress/integration/MegaMenuHover.feature
@@ -2,7 +2,7 @@ Feature: Analytics events raised by hovering over mega menu in desktop, clicking
 
 
     Scenario Outline: Hover over mega menu home page
-        Given user is navigating to "/" with added wait
+        Given user is navigating to "/"
         When user is hovering over mega menu with href "/about-cancer"
         And page click request is sent
         Then the following parameters should be captured
@@ -29,7 +29,7 @@ Feature: Analytics events raised by hovering over mega menu in desktop, clicking
             | /about-nci       |
 
     Scenario Outline: Hover over mega menu Spanish home page
-        Given user is navigating to "/espanol" with added wait
+        Given user is navigating to "/espanol"
         When user is hovering over mega menu with href "<href>"
         And page click request is sent
         Then the following parameters should be captured
@@ -57,7 +57,7 @@ Feature: Analytics events raised by hovering over mega menu in desktop, clicking
 
 
     Scenario: Mobile mega menu reveal on Home Page
-        Given user is navigating to "/" with added wait
+        Given user is navigating to "/"
         When screen breakpoint is set to "mobile"
         When user is clicking on mega menu button
         And page click request is sent
@@ -77,7 +77,7 @@ Feature: Analytics events raised by hovering over mega menu in desktop, clicking
 
 
     Scenario: Mobile mega menu reveal on Spanish home page
-        Given user is navigating to "/espanol" with added wait
+        Given user is navigating to "/espanol"
         When screen breakpoint is set to "mobile"
         When user is clicking on mega menu button
         And page click request is sent
@@ -98,7 +98,7 @@ Feature: Analytics events raised by hovering over mega menu in desktop, clicking
 
 
     Scenario: Tablet mega menu reveal is behaving as mobile breakpoint on home page
-        Given user is navigating to "/" with added wait
+        Given user is navigating to "/"
         When screen breakpoint is set to "tablet"
         When user is clicking on mega menu button
         And page click request is sent
@@ -120,7 +120,7 @@ Feature: Analytics events raised by hovering over mega menu in desktop, clicking
 
 
     Scenario: Tablet mega menu reveal is behaving as mobile breakpoint on Spanish home page
-        Given user is navigating to "/espanol" with added wait
+        Given user is navigating to "/espanol"
         When screen breakpoint is set to "tablet"
         When user is clicking on mega menu button
         And page click request is sent
