+++ cypress/utils/index.js
@@ -1,3 +1,4 @@
 export {convertBeacon} from "./analyticsBeacon";
 export {getClickBeacon} from "./getClickBeacon";
 export {getLoadBeacon} from "./getLoadBeacon";
+export {InitializeAnalyticsInterceptingImage} from "./analyticsInterceptingImage";
